refactor: recent transaction to compose
