+++ app/src/main/java/org/mifos/mobile/ui/recent_transaction/RecentTransaction.kt
@@ -1,0 +1,10 @@
+package org.mifos.mobile.ui.recent_transaction
+
+data class RecentTransaction(
+
+    val value: String? = null,
+
+    val amount: String? = null,
+
+    val date: String = ""
+)
