+++ app/src/main/java/org/mifos/mobile/ui/recent_transaction/RecentTransactionItem.kt
@@ -1,0 +1,51 @@
+package org.mifos.mobile.ui.recent_transaction
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import org.mifos.mobile.R
+
+/**
+ * @author pratyush
+ * @since 03/01/2024
+ */
+
+@Composable
+fun RecentTransactionItem(amount: String, date: String, value: String) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(8.dp)
+    ) {
+        Image(
+            modifier = Modifier.size(39.dp),
+            painter = painterResource(id = R.drawable.ic_local_atm_black_24dp),
+            contentDescription = null
+        )
+
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(start = 8.dp)
+        ) {
+            Text(text = value, color = Color.White)
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween
+            ) {
+                Text(text = amount, color = Color.White)
+                Text(text = date, color = Color.White)
+            }
+        }
+    }
+}
