+++ app/src/main/java/org/mifos/mobile/ui/recent_transaction/RecentTransactionViewModel.kt
@@ -1,4 +1,4 @@
-package org.mifos.mobile.viewModels
+package org.mifos.mobile.ui.recent_transaction
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
