+++ app/src/test/java/org/mifos/mobile/viewModels/RecentTransactionViewModelTest.kt
@@ -14,6 +14,7 @@ import org.mifos.mobile.models.Transaction
 import org.mifos.mobile.models.client.Currency
 import org.mifos.mobile.models.client.Type
 import org.mifos.mobile.repositories.RecentTransactionRepository
+import org.mifos.mobile.ui.recent_transaction.RecentTransactionViewModel
 import org.mifos.mobile.util.RxSchedulersOverrideRule
 import org.mifos.mobile.utils.RecentTransactionUiState
 import org.mockito.Mock
